Move Commit/Transaction into separate files
